Clean up integration review findings
