+++ internal/cli/issue_vcs_branch.go
@@ -98,9 +98,7 @@ func addIssueVCSBranchAttachmentsCommand(ctx context.Context, root *cobra.Comman
 			list.Attachments = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.AttachmentSummary) error {
-			return writeAttachment(command, options, item)
-		},
+		writeAttachment,
 		func(list client.AttachmentList) []client.AttachmentSummary {
 			return list.Attachments
 		},
@@ -146,9 +144,7 @@ func addIssueVCSBranchChildrenCommand(ctx context.Context, root *cobra.Command, 
 			list.Issues = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.IssueSummary) error {
-			return writeIssue(command, options, item)
-		},
+		writeIssue,
 		func(list client.IssueList) []client.IssueSummary {
 			return list.Issues
 		},
@@ -174,9 +170,7 @@ func addIssueVCSBranchDocumentsCommand(ctx context.Context, root *cobra.Command,
 			list.Documents = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.DocumentSummary) error {
-			return writeDocument(command, options, item)
-		},
+		writeDocument,
 		func(list client.DocumentList) []client.DocumentSummary {
 			return list.Documents
 		},
@@ -202,9 +196,7 @@ func addIssueVCSBranchFormerAttachmentsCommand(ctx context.Context, root *cobra.
 			list.Attachments = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.AttachmentSummary) error {
-			return writeAttachment(command, options, item)
-		},
+		writeAttachment,
 		func(list client.AttachmentList) []client.AttachmentSummary {
 			return list.Attachments
 		},
@@ -230,9 +222,7 @@ func addIssueVCSBranchHistoryCommand(ctx context.Context, root *cobra.Command, o
 			list.History = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.IssueHistorySummary) error {
-			return writeIssueHistory(command, options, item)
-		},
+		writeIssueHistory,
 		func(list client.IssueHistoryList) []client.IssueHistorySummary {
 			return list.History
 		},
@@ -277,9 +267,7 @@ func addIssueVCSBranchLabelsCommand(ctx context.Context, root *cobra.Command, op
 			list.Labels = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.LabelSummary) error {
-			return writeLabel(command, options, item)
-		},
+		writeLabel,
 		func(list client.LabelList) []client.LabelSummary {
 			return list.Labels
 		},
@@ -338,9 +326,7 @@ func addIssueVCSBranchReleasesCommand(ctx context.Context, root *cobra.Command, 
 			list.Releases = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.ReleaseSummary) error {
-			return writeRelease(command, options, item)
-		},
+		writeRelease,
 		func(list client.ReleaseList) []client.ReleaseSummary {
 			return list.Releases
 		},
@@ -386,9 +372,7 @@ func addIssueVCSBranchStateHistoryCommand(ctx context.Context, root *cobra.Comma
 			list.Spans = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.IssueStateSpanSummary) error {
-			return writeIssueStateSpan(command, options, item)
-		},
+		writeIssueStateSpan,
 		func(list client.IssueStateHistoryList) []client.IssueStateSpanSummary {
 			return list.Spans
 		},
@@ -414,9 +398,7 @@ func addIssueVCSBranchSubscribersCommand(ctx context.Context, root *cobra.Comman
 			list.Users = items
 			return list, err
 		},
-		func(command *cobra.Command, item client.UserSummary) error {
-			return writeUser(command, options, item)
-		},
+		writeUser,
 		func(list client.UserList) []client.UserSummary {
 			return list.Users
 		},
